Move side panel tests
